refactored TagButton
